refactor: move `SIP` generation to separate package
This is required for generating `Submission Information Package`(SIP)
for different types of submissions. For example, acquisitions.

No logic changes.
