+++ internal/send/send.go
@@ -10,20 +10,11 @@ import (
 	"time"
 
 	kafkaHelpers "hermetic/internal/kafka"
+	"hermetic/internal/submission_information_package"
 
 	"github.com/allegro/bigcache/v3"
-	"github.com/google/uuid"
 	"github.com/segmentio/kafka-go"
 )
-
-type TransferSubmissionInformationPackage struct {
-	Date            string `json:"date"`
-	ContentCategory string `json:"contentCategory"`
-	ContentType     string `json:"contentType"`
-	Identifier      string `json:"identifier"`
-	Urn             string `json:"urn"`
-	Path            string `json:"path"`
-}
 
 type offsets struct {
 	first int64
@@ -56,7 +47,7 @@ func getFirstAndLastOffsets(kafkaEndpoints []string, transferTopicName string) (
 	return offsets{first: firstOffset, last: lastOffset}, nil
 }
 
-func readLatestMessages(kafkaEndpoints []string, transferTopicName string) ([]TransferSubmissionInformationPackage, error) {
+func readLatestMessages(kafkaEndpoints []string, transferTopicName string) ([]submission_information_package.SubmissionInformationPackage, error) {
 	messageReader := kafkaHelpers.MessageReader{
 		Reader: kafka.NewReader(kafka.ReaderConfig{
 			Brokers: kafkaEndpoints,
@@ -72,7 +63,7 @@ func readLatestMessages(kafkaEndpoints []string, transferTopicName string) ([]Tr
 	}
 	readTimeout := 10 * time.Second
 
-	var messages []TransferSubmissionInformationPackage
+	var messages []submission_information_package.SubmissionInformationPackage
 
 	for offsetToReadFrom := offsets.first; offsetToReadFrom < offsets.last; offsetToReadFrom++ {
 		fmt.Printf("Reading message at offset '%d'\n", offsetToReadFrom)
@@ -93,7 +84,7 @@ func readLatestMessages(kafkaEndpoints []string, transferTopicName string) ([]Tr
 			fmt.Printf("Message at offset '%d' is nil, skipping offset\n", offsetToReadFrom)
 			continue
 		}
-		var transferSubmissionInformationPackage TransferSubmissionInformationPackage
+		var transferSubmissionInformationPackage submission_information_package.SubmissionInformationPackage
 
 		err = json.Unmarshal(message.Value, &transferSubmissionInformationPackage)
 		if err != nil {
@@ -111,8 +102,8 @@ func readLatestMessages(kafkaEndpoints []string, transferTopicName string) ([]Tr
 	return messages, nil
 }
 
-func webArchiveRelevantMessages(messages []TransferSubmissionInformationPackage) ([]TransferSubmissionInformationPackage, error) {
-	var relevantMessages []TransferSubmissionInformationPackage
+func webArchiveRelevantMessages(messages []submission_information_package.SubmissionInformationPackage) ([]submission_information_package.SubmissionInformationPackage, error) {
+	var relevantMessages []submission_information_package.SubmissionInformationPackage
 	for _, message := range messages {
 		if message.ContentCategory == "nettarkiv" {
 			if message.ContentType != "warc" {
@@ -183,7 +174,7 @@ func PrepareAndSendSubmissionInformationPackage(kafkaEndpoints []string, transfe
 				return fmt.Errorf("found file '%s' in root path '%s', but expected only directories", path.Name(), rootPath)
 			}
 			fmt.Printf("Processing directory %s\n", destinationPath)
-			transferSubmissionInformationPackage := createSubmissionInformationPackage(destinationPath, directoryName)
+			transferSubmissionInformationPackage := submission_information_package.Create(destinationPath, directoryName)
 
 			kafkaMessage, err := json.Marshal(transferSubmissionInformationPackage)
 			if err != nil {
@@ -201,23 +192,5 @@ func PrepareAndSendSubmissionInformationPackage(kafkaEndpoints []string, transfe
 
 		}
 		time.Sleep(1 * time.Minute)
-	}
-}
-
-func createSubmissionInformationPackage(payloadPath string, payloadDirName string) TransferSubmissionInformationPackage {
-	date := time.Now().UTC().Format("2006-01-02T15:04:05.000")
-	contentCategory := "nettarkiv"
-	contentType := "warc"
-	commonPart := "no-nb_" + contentCategory + "_" + payloadDirName
-	identifier := commonPart + "_" + uuid.New().String()
-	urn := "URN:NBN:" + commonPart
-
-	return TransferSubmissionInformationPackage{
-		Date:            date,
-		ContentCategory: contentCategory,
-		ContentType:     contentType,
-		Identifier:      identifier,
-		Urn:             urn,
-		Path:            payloadPath,
 	}
 }
